requirements.txt
# Core LLM Framework
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langserve[all]>=0.3.0

# Backend & API
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0

# Document Processing
pymupdf==1.23.8
python-docx==1.1.0

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.2.2

# Frontend
gradio==4.14.0

# Evaluation & Data
pandas==2.1.4